# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sparnatural Query2Text, Text2Query & Reconciliation API
  version: 1.0.0
  description: |
    Cette API fournit trois fonctionnalités principales pour interagir avec Sparnatural AI:

    - **Query2Text** : Convertit une requête JSON structurée en une phrase lisible en langage naturel.
    - **Text2Query** : Génère une requête JSON (au format Sparnatural) à partir d’une question ou d’une demande en langage naturel.
    - **Reconciliation** : Rechercher des URI RDF pertinentes à partir d’un label (utile pour la réconciliation ou l’autocomplétion).

    Chaque route peut être personnalisée par projet (`projectKey`) et supporte plusieurs langues.
  tags:
    - name: Query2Text
      description: Génération de texte en langage naturel depuis une requête JSON
    - name: Text2Query
      description: Génération d'une requête JSON à partir d'une phrase
    - name: Reconciliation
      description: Recherche d'URI RDF pour un label donné

paths:
  /api/v1/{projectKey}/query2text:
    get:
      summary: Convertit une requête JSON en texte en langage naturel
      tags: [Query2Text]
      description: >
        Fournit une version textuelle (langage naturel) d'une requête structurée Sparnatural. 
        Cette conversion peut être adaptée selon le projet (`projectKey`) et la langue (`lang`).
      parameters:
        - in: path
          name: projectKey
          required: true
          schema:
            type: string
            example: dbpedia-en
        - in: query
          name: query
          required: true
          schema:
            type: string
        - name: lang
          in: query
          required: false
          description: Code de langue ISO 639-1 (par défaut "en"). Permet d'obtenir le texte dans une langue donnée.
          schema:
            type: string
            default: en
            enum: [en, fr]
            example: fr
      responses:
        "200":
          description: Résumé généré
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
              examples:
                success:
                  summary: Exemple de résumé
                  value:
                    text: "Afficher toutes les œuvres d’art exposées dans un musée."
        "400":
          description: Requête invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid query parameter"
        "500":
          description: Erreur interne
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

  /api/v1/{projectKey}/text2query:
    get:
      summary: Convertit une demande en langage naturel en requête JSON (Sparnatural)
      tags: [Text2Query]
      description: >
        Fournit une requête structurée au format Sparnatural à partir d’une phrase écrite en langage naturel.
      parameters:
        - in: path
          name: projectKey
          required: true
          schema:
            type: string
            example: dbpedia-en
        - in: query
          name: text
          required: true
          description: Demande ou requête en langage naturel.
          schema:
            type: string
            example: "Donne-moi toutes les personnes nées en France"
      responses:
        "200":
          description: Requête JSON générée
          content:
            application/json:
              schema:
                type: object
              examples:
                success:
                  summary: Exemple de requête générée
                  value:
                    query: object
        "400":
          description: Requête invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid query parameter"
        "500":
          description: Erreur interne
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

  /api/v1/{projectKey}/ping:
    get:
      summary: Ping the project
      tags: [Ping]
      description: >
        Test whether a specific project is reachable and its config was properly read
      parameters:
        - in: path
          name: projectKey
          required: true
          schema:
            type: string
            example: dbpedia-en
      responses:
        "200":
          description: Project reachable
          content:
            application/json:
              schema:
                type: string
              examples:
                success:
                  summary: Exemple de requête générée
                  value:
                    query: object
        "404":
          description: Unknown project key
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unknown project key"

  /api/v1/{projectKey}/prompt-t2q:
    get:
      summary: Generate the Text2Query prompt for a project
      tags: [T2QPromptGenerator]
      description: >
        Generates the complete Text2Query prompt for a given project.
        The SHACL file is automatically read from the project configuration (config.yaml)
        using the `projectKey`. The returned prompt can be used directly in a text2query agent.
      parameters:
        - in: path
          name: projectKey
          required: true
          description: The project identifier (e.g. "dbpedia-en", "demo-ep").
          schema:
            type: string
            example: dbpedia-en
      responses:
        "200":
          description: Prompt generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
                    description: The full text2query prompt ready to be used in an agent.
              examples:
                success:
                  summary: Example response
                  value:
                    prompt: "Role: Semantic Query Builder Assistant\n\nObjective: Translate a user's natural language request..."
        "404":
          description: Project not found or no SHACL file configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                unknownProject:
                  summary: Unknown project
                  value:
                    error: "Unknown project: 'my-project'"
                noShacl:
                  summary: No SHACL file configured
                  value:
                    error: "No SHACL file configured for project 'my-project'"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error generating prompt"
  /api/v1/{projectKey}/prompt-q2t:
    get:
      summary: Generate the Query2Text prompt for a project
      tags: [Q2TPromptGenerator]
      description: >
        Generates the complete Query2Text prompt for a given project.
        The SHACL file is automatically read from the project configuration (config.yaml)
        using the `projectKey`. The returned prompt can be used directly in a query2text agent.
      parameters:
        - in: path
          name: projectKey
          required: true
          description: The project identifier (e.g. "dbpedia-en", "demo-ep").
          schema:
            type: string
            example: dbpedia-en
      responses:
        "200":
          description: Prompt generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  prompt:
                    type: string
                    description: The full query2text prompt ready to be used in an agent.
              examples:
                success:
                  summary: Example response
                  value:
                    prompt: "Role: Semantic Query Assistant\n\nObjective: Translate a structured JSON query into a clear, concise, and natural-sounding sentence in English or French depending on the query's metadata language.\n\n---"
        "404":
          description: Project not found or no SHACL file configured
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                unknownProject:
                  summary: Unknown project
                  value:
                    error: "Unknown project: 'my-project'"
                noShacl:
                  summary: No SHACL file configured
                  value:
                    error: "No SHACL file configured for project 'my-project'"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Error generating prompt"

  /api/v1/{projectKey}/reconciliation:
    post:
      summary: Recherche d'URI à partir de plusieurs termes (format Reconciliation W3C)
      tags: [Reconciliation]
      description: >
        Permet de rechercher des URI pour plusieurs termes en une seule requête,
        selon la spécification W3C Reconciliation (https://www.w3.org/community/reports/reconciliation/).
      parameters:
        - in: path
          name: projectKey
          required: true
          schema:
            type: string
            example: dbpedia-en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: object
                required:
                  - query
                properties:
                  query:
                    type: string
                    description: Terme à rechercher
                    example: "France"
            example:
              q1:
                query: "France"
              q2:
                query: "Algeria"

      responses:
        "200":
          description: Résultats de recherche URI pour chaque requête
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
                  properties:
                    result:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: string
                            description: URI trouvée
                          name:
                            type: string
                            description: Terme recherché
                          score:
                            type: integer
                            description: Score de correspondance (100 = parfait)
                          match:
                            type: boolean
                            description: Indique si c’est un match certain
              examples:
                success:
                  summary: Exemple de réponse
                  value:
                    q1:
                      result:
                        - id: "http://dbpedia.org/resource/France"
                          name: "France"
                          score: 100
                          match: true
                    q2:
                      result:
                        - id: "http://dbpedia.org/resource/Algeria"
                          name: "Algeria"
                          score: 100
                          match: true
        "400":
          description: Requête invalide
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid input JSON format"
        "500":
          description: Erreur interne du serveur
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"